# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards Management API
  description: API for managing bank cards with admin functionality
  contact:
    name: API Support
    email: api@example.com
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
...
  /transactions/transfers:
    patch:
      tags:
        - Transactions Controller
      summary: Transfer funds between cards
      description: Transfer funds between two cards owned by the same authenticated
        user. Both cards must belong to the user.
      operationId: transfer
      requestBody:
        description: Transaction request details
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequest"
            example:
              fromCard: "1234567812345678"
              toCard: "8765432187654321"
              amount: 100.5
        required: true
      responses:
        "200":
          description: Transfer completed successfully
          content:
            application/json:
              schema:
                type: string
                enum:
                  - 100 CONTINUE
                  - 101 SWITCHING_PROTOCOLS
                  - 102 PROCESSING
                  - 103 EARLY_HINTS
                  - 103 CHECKPOINT
                  - 200 OK
                  - 201 CREATED
                  - 202 ACCEPTED
                  - 203 NON_AUTHORITATIVE_INFORMATION
                  - 204 NO_CONTENT
                  - 205 RESET_CONTENT
                  - 206 PARTIAL_CONTENT
                  - 207 MULTI_STATUS
                  - 208 ALREADY_REPORTED
                  - 226 IM_USED
                  - 300 MULTIPLE_CHOICES
                  - 301 MOVED_PERMANENTLY
                  - 302 FOUND
                  - 302 MOVED_TEMPORARILY
                  - 303 SEE_OTHER
                  - 304 NOT_MODIFIED
                  - 305 USE_PROXY
                  - 307 TEMPORARY_REDIRECT
                  - 308 PERMANENT_REDIRECT
                  - 400 BAD_REQUEST
                  - 401 UNAUTHORIZED
                  - 402 PAYMENT_REQUIRED
                  - 403 FORBIDDEN
                  - 404 NOT_FOUND
                  - 405 METHOD_NOT_ALLOWED
                  - 406 NOT_ACCEPTABLE
                  - 407 PROXY_AUTHENTICATION_REQUIRED
                  - 408 REQUEST_TIMEOUT
                  - 409 CONFLICT
                  - 410 GONE
                  - 411 LENGTH_REQUIRED
                  - 412 PRECONDITION_FAILED
                  - 413 PAYLOAD_TOO_LARGE
                  - 413 REQUEST_ENTITY_TOO_LARGE
                  - 414 URI_TOO_LONG
                  - 414 REQUEST_URI_TOO_LONG
                  - 415 UNSUPPORTED_MEDIA_TYPE
                  - 416 REQUESTED_RANGE_NOT_SATISFIABLE
                  - 417 EXPECTATION_FAILED
                  - 418 I_AM_A_TEAPOT
                  - 419 INSUFFICIENT_SPACE_ON_RESOURCE
                  - 420 METHOD_FAILURE
                  - 421 DESTINATION_LOCKED
                  - 422 UNPROCESSABLE_ENTITY
                  - 423 LOCKED
                  - 424 FAILED_DEPENDENCY
                  - 425 TOO_EARLY
                  - 426 UPGRADE_REQUIRED
                  - 428 PRECONDITION_REQUIRED
                  - 429 TOO_MANY_REQUESTS
                  - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
                  - 451 UNAVAILABLE_FOR_LEGAL_REASONS
                  - 500 INTERNAL_SERVER_ERROR
                  - 501 NOT_IMPLEMENTED
                  - 502 BAD_GATEWAY
                  - 503 SERVICE_UNAVAILABLE
                  - 504 GATEWAY_TIMEOUT
                  - 505 HTTP_VERSION_NOT_SUPPORTED
                  - 506 VARIANT_ALSO_NEGOTIATES
                  - 507 INSUFFICIENT_STORAGE
                  - 508 LOOP_DETECTED
                  - 509 BANDWIDTH_LIMIT_EXCEEDED
                  - 510 NOT_EXTENDED
                  - 511 NETWORK_AUTHENTICATION_REQUIRED
        "400":
          description: Invalid request parameters
          content:
            application/json:
              examples:
                Invalid amount:
                  description: Invalid amount
                  value:
                    error: Amount must have at most 2 decimal places
                    timestamp: "1757615790252"
                Same card transfer:
                  description: Same card transfer
                  value:
                    error: Transfer from the card to itself is not possible
                    timestamp: "1757615790252"
        "404":
          description: Card not found
          content:
            application/json:
              example:
                error: Card not found
                timestamp: "1757615790252"
        "403":
          description: Access denied - cards belong to different users
          content:
            application/json:
              example:
                error: You do not have permission to transfer with these cards
                timestamp: "1757615790252"
        "402":
          description: Insufficient funds
          content:
            application/json:
              example:
                error: Insufficient funds
                timestamp: "1757615790252"
...
components:
  schemas:
...
    TransactionRequest:
      type: object
      description: Transaction request DTO for fund transfers between cards
      properties:
        fromCard:
          type: string
          description: Source card number (16 digits)
          example: 1234567812345678
          maxLength: 16
          minLength: 16
          pattern: ^\d+$
        toCard:
          type: string
          description: Destination card number (16 digits)
          example: 8765432187654321
          maxLength: 16
          minLength: 16
          pattern: ^\d+$
        amount:
          type: number
          description: "Amount to transfer (minimum 0.01, maximum 2 decimal places)"
          example: 100.5
          minimum: 0.01
      required:
        - fromCard
        - toCard
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT